# data-service/src/main/resources/openapi/schema-service-api.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "schema-service-api",
    "description": "Schema Service API",
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "http://localhost:8092",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/v1/schemas/schema/{subject}/{version}": {
      "get": {
        "tags": [
          "schema-service"
        ],
        "operationId": "getSchemaBySubjectAndVersion",
        "parameters": [
          {
            "name": "subject",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetSchemaResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GetSchemaResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "SUCCESS",
              "FAILURE"
            ]
          },
          "message": {
            "type": "string"
          },
          "schema": {
            "type": "string"
          }
        }
      },
...
    }
  }
}